feat: [PLG-1569]: ci designing sub page
